Rename Loader to DefinitionLoader
